fix(shell): account for "combined" `SIGCHLD` signals
There is no queue for signals, so sometimes it is possible for multiple
occurrences of the same signal to get merged into one. We can see this
effect if we enter `sleep 3 | sleep 3 | sleep 3 | ...` enough times.
Hence, it is also possible for multiple `SIGCHLD` signals to combine.
We need to use a loop within the signal handler to consume multiple
zombie processes if that is the case.