fix(builtins, run): fix compilation failure with musl
`stdin`, `stdout` and `stderr` are macros (dictated by the C standard),
so they actually cannot be used as identifiers.

With glibc, the code still compiled fine (not exactly sure why, but
probably related to implementation details). On musl, however, the code
does not compile.

This commit fixes the source code so that it does not use those reserved
names as identifiers.